Implement account defaults, month-to-month navigation, and UI theme refinements.
This commit introduces significant improvements to account management, dashboard navigation, and the overall design system, including a new `Decimal` abstraction for precise financial calculations and a custom Material 3 theme.

Key changes:
- **Account Management:**
    - Implement "Default Account" functionality, allowing users to star a specific account.
    - Automatically select the default account when creating new transactions.
    - Update repositories and SQLDelight schemas to persist and retrieve the default account status.
- **Dashboard & Reporting:**
    - Add month-to-month navigation on the dashboard with a new `TopBarWithNavigation` component.
    - Support filtering dashboard and report data by specific accounts.
    - Enhance dashboard visuals with updated `compose-charts` configurations and more descriptive summary cards.
    - Map categories to recent transactions on the dashboard for better visibility.
- **Design System & Theming:**
    - Implement `CashFlowTheme` with custom light/dark color schemes and semantic finance colors (Income/Expense).
    - Support dynamic color on Android (Material You).
    - Introduce a centralized `DSFullscreenLoader` component and refactor screens to use it.
    - Standardize transaction amount colors across the app using `AppColors`.
- **Core Domain & Data:**
    - Introduce a `Decimal` expect/actual class to handle currency formatting and double conversions across platforms (using `BigDecimal` on JVM/Android).
    - Refactor all domain models (`Account`, `Transaction`, `Budget`, `RecurringRule`) to use `Decimal` for amount fields.
    - Update Data Mappers and SQLDelight repositories to handle the new `Decimal` type.
- **Housekeeping:**
    - Reorganize UI package structure, moving screens into a dedicated `ui.screens` subpackage.
    - Add utility for formatted date pattern strings across platforms.
    - Clean up unused platform abstraction boilerplate in `Platform.android.kt`.

+++ composeApp/src/commonMain/kotlin/com/github/dsbezerra/cashflow/data/mapper/BudgetMapper.kt
@@ -2,11 +2,12 @@ package com.github.dsbezerra.cashflow.data.mapper
 
 import com.github.dsbezerra.cashflow.db.Budget as BudgetEntity
 import com.github.dsbezerra.cashflow.domain.model.Budget
+import com.github.dsbezerra.cashflow.domain.model.toDecimal
 
 fun BudgetEntity.toDomain(): Budget = Budget(
     id = id,
     categoryId = categoryId,
-    amount = amount,
+    amount = amount.toDecimal(),
     period = period,
     startDate = startDate,
     isActive = isActive,
@@ -15,7 +16,7 @@ fun BudgetEntity.toDomain(): Budget = Budget(
 fun Budget.toEntity(): BudgetEntity = BudgetEntity(
     id = id,
     categoryId = categoryId,
-    amount = amount,
+    amount = amount.toDouble(),
     period = period,
     startDate = startDate,
     isActive = isActive,
